sets up server to connect to soundcloud api, adds querystring and node-fetch modules
