--- server/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "radiar-server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.13.3",
-    "spotify-web-api-node": "^2.2.0"
-  }
-}
